+++ R/BioDataScience1-package.R
@@ -5,7 +5,10 @@
 #' @docType package
 #' @name BioDataScience1-package
 #'
-#' @importFrom learndown config run learndownLearnrBanner learndownLearnrServer
+#' @importFrom learndown run learndownLearnrBanner learndownLearnrServer
+#' @importFrom BioDataScience config sign_in sign_out
+#' @importFrom shiny dialogViewer observeEvent selectInput stopApp runGadget
+#' @importFrom miniUI gadgetTitleBar miniContentPanel miniPage miniTitleBarButton miniTitleBarCancelButton
 NULL
 
 #@importFrom BioDataScience init